Basic obstacles set up
Added trampoline ladder with moving platform at the top that can transport the player across the gap and over to the flag. Players also have the option of bypassing the trampoline ladder and boldly attempting to leap across the gap using their unique abilities. Still need to fix the scaling bug that occurs when a player collides with the moving platform.